+++ src/main/java/utils/PackageConvertor.java
@@ -9,7 +9,7 @@ public class PackageConvertor {
     public static PackageInfo convertPackage(final String rawPackage) {
         final String[] info = rawPackage.split(SPACE);
         final String postalCode = info[1];
-        final Double weight = Double.parseDouble(info[0]);
+        final Float weight = Float.parseFloat(info[0]);
 
         return new PackageInfo(postalCode, weight);
     }
